Add flag to selectively skip UI initializations from gradle unit tests
The GradleProjectImporter performs some logic when a project has been
imported.  Some of that logic does not work from the limited context
of android unit tests (where we are not running a full IntelliJ
instance). In an earlier CL I turned off some of that handling, from
unit tests, but it turns out that that interferes with some existing
tests which are extending IdeaTestCase to run the IDE.

This changeset tweaks this by adding a flag, set by these limited
tests, to more selectively skip the UI initialization steps. This
isn't beautiful, but it gets the tests all passing again. Longer term
we should probably make the gradle-based tests all extend IdeaTestCase
instead, but that's going to require moving a bunch of test utility
code (such as all the code to initialize SDKs) such that it can be
accessed from both AndroidTestCase, which does not extend
IdeaTestCase, and AndroidGradleTestCase, which now should.

Change-Id: Id2ef822939dfbbbfd5455cf0d93a875831872e8a