Support non-rectangular shapes in drawable render preview
This CL ensures that the render preview handles XML drawables with non-rectangular
shapes, such as a <shape><corners android:radius="100dp" />..., correctly. In particular,
the image is painted with a transparent background (such that we don't insert the theme
background), and if necessary, we create a non-rectangular drop shadow rather than the
usual (and faster) rectangular drop shadow.

Change-Id: I77f5f4b69526ace5af0043af92d1eae4b917f597