Enabled building selected modules.
(By "building" we mean "making" and "compiling.")

Currently, Android Studio builds all the modules
in the project. Users have reported that they
only want to build certain modules. This CL
enables that.

The main issue with JPS is that it is impossible
to have a 'target builder' (in JPS terminology)
that knows about all the selected modules in
the IDE (JPS has no notion of an IDE.) The
most that we can get from JPS is to have one
'target builder' per module, and these target
builders won't know about each other, making
it impossible to have a single call to Gradle
passing the tasks to execute per selected
module (we would end with a call to Gradle
per module.)

To work around this, we have a back door: we
have an extension point in JPS where we can pass
JVM args to the JPS process (we requested that
some time ago.) In this CL we pass the names
of the selected modules (if any) to as JVM
parameters to the JPS process.

On the other hand, it is really easy to implement
building selected modules in the new build
strategy.

In addition, this CL:
- Fixes the "update" mechanism of the "Make
  Module(s)" and "Compile Module(s)".
- Makes it possible to compile a module
  if any of its files is selected.
- Unifies the way Gradle tasks to execute
  are calculated (for both JPS and
  direct Gradle invocation,) fixing bugs
  in both build approaches.
- Does not call "assemble" on pure Java
  libraries  when generating sources.
- Cleans up code.
- Adds more tests.

Fixes
https://code.google.com/p/android/issues/detail?id=59915

Change-Id: I4f24fbfd4773d37c0276ee7b6815d85f6b312fa0