Layout rendering: Convert class files on the fly
This changeset adds a render class loader which converts classes on the fly,
such that if you have compiled your project with for example javac from 1.7,
and you are running on 1.6, it can load the R class and most typical custom
views and run them, instead of throwing UnsupportedClassVersionErrors as
before.

This is done by loading the classes in and rewriting them with ASM, then
loading the resulting bytes instead of the original classfile contents.
This may not work in all cases, but it works for not only the rather simple
code in most R classes, but even for custom views with multi catch statements
etc.

This CL also makes the class file search method more efficient, and improves
the error messages for cases where rewriting doesn't work (by for example
tracking the required version of class files in a new exception that is
passed on to the error handler.)

This helps with CL
61669: Layout Preview - Target 1.7 - unsupported class version

Change-Id: If8f21a20939bff330943da8ab42143742a3f85fa