Refactoring of BuildFileKey type system.
This CL doesn't introduce any (intended) change in functionality.
It takes the Type enum that used to live inside BuildFileKey and
raises it to a top-level type. Functionality that used to live
throughout the code in the form of switch statements on that type
has by and large been converted to method calls into this type
class, making client code more streamlined. The parsing code
in BuildFileKey has been simplified. It mostly delegates to the
type class to do parsing/output. Previously it used to have
different code paths for keys that have closure types vs. keys
that have simple types; this has been merged together.

Change-Id: I1dca4fc47dd844f026ca9bac4988b7e69ae70924