More work on parsing Java code in projects
This is the first cut of the Navigation Editor that gets all of the transitions of
SimpleMail directly from the project.

Also added the "Transient" annotation for the XML archiving scheme and marked
the transitions of the NavigationModel as transient.

The XML archive no longer needs to be distributed with the app but is retained for
design-time information like where screens are placed in the Navigation Model
diagrams.

Change-Id: If6c635e19b83cb79d1f32011db1bcd88724c308f