Now we ask user to choose the path of Android SDK.
Every time a project is synced with Gradle (e.g. when
importing a project) and if the project's
local.properties points to an Android SDK different
than the one specified in the IDE, we ask the user to
choose the SDK to use.

If the user chooses the SDK path in local.properties,
the IDE is updated to use the same path. If the user
chooses the IDE's SDK path, the local.properties file
is updated with the new path.

The purpose of this CL is to have build consistent
between IDE and command line.

Change-Id: Ibcaf53128da782fe9322b97f5e982431654144df