Fix editor/project initialization bug
This CL fixes the bug where creating a new project would open the layout editor,
but it would not have Design and Text tabs, so you couldn't for example switch
to the layout editor.

The reason for this was that the template files are opened "too early"; at the
point projectInitialized is called, the AndroidFacet is still missing from
the project.

The fix is to recognize this scenario and use the StartupManager to schedule
the files to be opened as a post startup activity.

Change-Id: If9088510adc20c49ea270be567938b5dec0c7ab3