Now we warn users about 'auto-import' and offer to disable it.
Due to the fact that Gradle task execution cannot be cancelled,
the 'auto-import' slows down the IDE when editing build.gradle
files.

This CLs shows a notification when this setting is enabled, and
offers a link to the bug report and a way to turned this feature
off using a single click.

Change-Id: I648ba3b817444db948f9f861f65c9f5e8bc38259