Distinguish between module, project and app resources
Until now, we had two types of local resources:
- the resources in a module
- the merged resources from the module, any modules it depends on, and
  any libraries it depends on.

The latter is used for example during rendering or when resolving
resource references in the editor.

However, we sometimes need to get a merged view of the resources
*without* the libraries. For example, when showing the locales in the
project, we want to base this on the locales present in the user's
modules, but not translations provided in any of the libraries used by
the app.

This changeset adds that. We now have the following hierarchy:

"AppResourceRepository" provides a merged view of all the local
(e.g. non-framework resources).  It merges repositories for each of
the AAR libraries with the project resources.

"ProjectResourceRepository" provides a merged view of all the
resources in the modules in the project.

"ModuleResources" provides a merged view of all the resource folders
in the module.

As before, the ResourceFolderRepository and the FileResourceRepository
provide lookup of resources in a specific PSI folder or java.io.File
folder respectively.

To do this, the previous ModuleSetRepository was cloned; one half is
now the ProjectResourceRepository (handling module resources) and the
other half is the AppResourceRepository (combining the previous
ProjectResourceRepository and each of the libraries). In addition, the
former "ProjectResources" was renamed to LocalResourceRepository, to
avoid confusion with the term "project resources" now meaning
resources in the project, not including libraries (which are now
referred to as app resources).

Finally, this CL cleans up usages such that for example locale lookup
uses project resources rather than app resources, there are explicit
named lookup methods now rather than fetching different versions of
the resources with based on several boolean parameters in the lookup
method, etc.

(Also fixes the RTL refactoring to not attempt fixing layouts in AAR
exploded bundle directories.)

Change-Id: I079afe31e97d1ba5587d207a990762904fdc355c