Better package selection for New Android Component
This change adds better logic for finding the package that the
user is targeting. It also adds a default icon for templates
that don't have an icon of their own.

It also enables the New Activity and New Android Component
actions everywhere, defaulting to the app package if no
java source folder is selected at the time of the action
invocation.
Change-Id: I3baf3f53c9d0dda82b3d8380891b72e16a990806