Fixed compiler outdir dirs for Java library modules.
For Java library modules, the compiler output path and test
output path weren't being set.

It is still not clear to me why this stopped working. In this
CL I set the compiler output path the same way as IDEA's
ModuleDataService does. The changes don't look any better to
me but seems to solve the issue. The only good thing for sure
is that we now check for null when obtaining the
CompilerModuleExtension for a module.

Change-Id: I266a86936eca69a0e385172200274cde4381deb6