Generate R classes for AAR custom views on the fly
This changeset addresses issue 62369: UI designer can not resolve
resources of android.support.v7.widget.GridLayout

The root problem is that AAR libraries ship with a classes.jar file
which includes the custom view, but NOT the R class it was compiled
with. This results in NoClassDefFoundError errors (different from
ClassNotFoundException) when you open a layout where the application
has not been compiled.

This changeset adds hooks into the render class loader such that it
detects this scenario. We already know all the resources in the AAR
library (since we parse these for layoutlib anyway), so all we need to
do is manufacture a class which represents the R class for these
resources. This changeset also adds a unit test which verifies that
the generated bytecode is valid and correct.

Change-Id: Ib838c5cc72ebfbf289207d1f2af0e8157c59607b