Fix incremental inspections in non-primary manifest files
The code which runs inspections in the editor background would
only check manifest files if the manifest file was identical to
the primary manifest file in the project. That worked in old
projects, but in Gradle projects there can be many manifest files.
Instead of checking that the files are identical, run inspections
on all XML files that have the specific Android manifest name
(AndroidManifest.xml).

Change-Id: If88642563f2f34386798894fd3c017e89f00ebb2