Clean up RenderErrorPanelTest
This changeset tweaks the RenderErrorPanelTest in a few ways to try to
narrow down the failures seen when run on one of the build server machines.

First, it uses the test platform property ($ADT_TEST_PLATFORM) to force
a specific layoutlib when rendering. Second, it performs a manual diff
of the before and after HTML strings and dumps these to standard out if
they differ (since the junit report diff is very difficult to make sense
of), and finally it cleans up the code a bit such that a lot of the setup
code can be reused.

Change-Id: Ie2c167394c80b64b6a49c015de3a122d5ee6be98