Handle manifest XML errors in the error parser
This changeset updates the Gradle error parser to handle error
messages in merged manifest files such that it points back to
the original sources instead.

It also fixes a couple of other issues in the parser:

- Fix document caching such that it refreshes the cached document if
  the underlying file has changed

- Improve the code which tries to find the original cause error
  corresponding to the final Gradle error. It normally searches for an
  error message that matches the first quoted error output line, but
  in the case of the aapt runner, we should use the last line instead.

Also fix bug in the new quickfix for using mock locations in non-debug
contexts: If the directory already exists, don't try to create it.

Change-Id: I9ee709c82bc8e7685929bf6ac2c375c4ab555d57