Add render diagnostics when running on older JDK7 versions
Issue 63279 seems to only happen on older versions on JDK 7. This CL detects when
a temp file creation fails with a security exception, and checks whether the current
JDK is older than a recent built, and if so emits a tip that this might be the
cause.

(This has to be done here rather than on the security manager side, because JDK 7's
code to create temp files will catch the security exception thrown by the security
manager and rethrow its own exception in order to hide the file name, so inserting
an error message there doesn't help.)

Change-Id: I4321a90eccbe7b72be067147a60d4a533f20b43a