Update resource folders and lint Studio integration for 0.7.0 buildsystem
Include sources from the getMultiFlavorSourceProvider and from the
getVariantSourceProvider source sets. Clean up code to use the source
providers directly (no need to go from source provider to idea source
provider and back to file in the lint bridge), and finally fix the
getBuildSdk and getProguard methods to use data from the model.

Change-Id: I9865d79e52a454dc6a3ac28bc2a252644dd52449