Cleanup/Test ConfigureAndroidModuleStep
This change provides tests for the configuration page for a new module
(the first page of the NPW, and the second of the New Module Wizard).
It also adds a space between the application title and the autogenerated
name (My Project 1 as opposed to My Project1), and some small cleanups
to the step code to allow for easier testing.

This test also fixes the NPW tests.

Change-Id: I2a31dfd604ac4e280877b0a0e3d88d5a91a57f51