Setting default SDK does not update "SDK Manager" action.
The bug is as follows:
In a fresh installation of Android Studio, without an Android SDK,
the "SDK Manager" action in the "Welcome" page is disabled. After
setting the default SDK from the "Project Structure" dialog, the
"SDK Manager" is never updated and it's always disabled, even though
there is an Android SDK this time.

Change-Id: Ia2de2efe8f8591e2e80dd66143e4366280ab8891