Fixed creation of temp module when project import fails.
The problem is that we are using the name of the project + ".iml"
to create the tmp module, but after fixing the project and syncing,
the real top-level module uses the name of the project's root
folder instead, resulting in 2 .iml files, which is confusing.

For example, when creating project "My Application 2", the temp
module file will be "My Application 2.iml" but it should be
"MyApplication2.iml". This CL fixes that.

Change-Id: I079ce36980802a773212cdf68fe3674451589f0f