Now we sync Java lang level after executing a build.
The problem is the following:
When changing the Java language level (e.g from "1.6" to
"1.7) we used to update the project's language level in the
middle of the "sync with Gradle" process. By the time the
dialog to reload the project showed up, the project was
already building. Selecting "reload project" resulted in
the dialog "Cannot stop Gradle build" (or something
similar,) which is a really awful user experience.

This CL fixes this by updating the Java lang level
*after* the project build is finished. We cannot get
away from the "reload project" dialog (it is baked in
IDEA,) but at least it is a lot less confusing now.

Fixes:
https://code.google.com/p/android/issues/detail?id=63622

Change-Id: I3dd17614bb8bef1c2e1f26bf43e992ef591da578