Teach Gradle about Android plugin's DSL
This CL adds a resolve contributor that helps the Gradle
plugin to resolve custom elements defined by the Android
Gradle plugin. Specifically, it maps the following identifiers
in the gradle build script to specific classes inside the
Android Gradle plugin:

 android to com.android.build.gradle.AppExtension
 defaultConfig to com.android.builder.DefaultProductFlavor

As of now the only feature this provides is that you can
cross navigate from the DSL to the plugin's class files.
There are a few enhancements that need to come from IntelliJ
that will make this even better. Specifically:
 - It can't use the sources right now because the Android Gradle
   plugin's sources are in Groovy, and IJ only parses Java
   sources.jar files. As a result, Go to definition jumps
   to the classes instead of the sources.
 - The Javadocs for the Android Gradle plugin aren't downloaded
   automatically. Users can explicitly request them to be
   downloaded by using a custom task like:
       apply plugin: 'idea'

       idea {
           module {
               downloadJavadoc = true
           }
       }
   Eventually however, IJ will provide settings in the IDE
   rather than having these extensions required in the build
   scripts. Once that is done (or if users add the above snippet),
   Javadoc for these DSL elements will be provided.

Change-Id: I193d48d949d0eb120ddcd2877bdfe37e57eafcfa