ddms panel: Add heap dump action
This CL adds the "dump heap" action to the DDMS panel.
When invoked, it instructs the selected client's VM to intiate
a heap dump. When the heap dump file is received, it is then
saved to disk, with an optional step to run "hprof-conv" on it
to convert it to the standard HPROF format.

Change-Id: Ic46a3d22e6d04e6a755aa0cb1b56e9a224711f6c