Deal with unparseable content in build.gradle files
In the Project Structure editors, recognize and handle unparseable
content, which can be comments, unexpected method calls or statements,
or extraneous Groovy such as added closures on familiar method
calls.

For repositories and dependencies, this CL creates abstractions to
allow us to remember the original source for those objects and
write it out verbatim when we rebuild those objects in Groovy.
For other object types, we need to track individual key entries that
have unparseable values and know to not stomp on them when writing
values back out.

This CL dismantles the canParseValue mechanism that was originally
built into BuildFileKey, which was designed with this need in mind,
but like many anticipatory pieces of infrastructure, turned
out to be unsuited to the task when it came time to hook it up.

Bug: 63148
Change-Id: I57e37049adb8df3e84edd9f4826f15e91afe9a35