Consolidate compiled resource id code in app resource repository
For layoutlib purposes we need to keep track of the exact id's to use
for each R field; we're notified of this upon completion of
compilation, and for new id's (prior to compilation) we assign them
dynamically.

This was handled in the LocalResourceRepository, and some other
classes were delegating and merging id's in the repositories which
merge values from other nested repositories.

However, for layoutlib we *always* use the app resource repository, so
migrate all the code to that single instance and get rid of the
delegation code.

Change-Id: Ib8d2d9bd28e891aa3944d96aa662cb5c36f59da5