Change the way in which make before run task is disabled.
Android Studio prefers the use of "Gradle-aware make" as opposed
to the default make before run task for both the application run
and test run configurations.

This CL simply changes the way in which the default make is disabled.
Rather than explicitly enabling only Gradle-aware make, we explicitly
disable just the default make.

This allows other plugins to selectively enable their own pre-run
tasks if necessary.

This CL also includes a refactoring that to remove code that is
common to both the Android Run and Test Run Configuration types.

Change-Id: I29ecbc18d56f543868adeb38b7df637e73e33c9c