Fixed automatic fix of gradle wrapper and resync.
When importing a project using the wrong version of Gradle
in the wrapper, we show a balloon notification where we offer
to automatically fix the wrapper and resync the project.

The issue is that we were fixing the wrapper, but when
resyncing the projet, we were not ensuring that the wrapper
is being used for the resync.

Fixes issue
https://code.google.com/p/android/issues/detail?id=63928

Change-Id: Iaa2d9db9a2b2cc6f01fec06a2929aaa4e88700b3