64239: string resources are not updated correctly in UI preview
A correct XML resource value file will contain uniquely named
items. If you repeat the definition of an item, the resource merger
will fail with a build error. Therefore, the ResourceFile class stores
items in a map.

However, in the IDE, the user may *temporarily* or *accidentally* have
multiple items with the same name. For example, when adding a new
string the user may simply duplicate the current line, then replace
the value and name fields.

This would break the incremental update handler for resource items,
since the map values for aliased resource items would be replaced and
therefore not updated. This changeset extends the IDE's custom
resource file class to keep a map of duplicate keys in files, keep
these up to date across item add and delete operations, and to make
the find method search not just the normal map but the duplicated
items as well.

Change-Id: Iaad52275838e448281ff488d5bf93fb24d77ece1