64653: lint quickfix: replace with suggested characters doesn't work
The lint error message was HTML escaped by the inspections integration code
such that it could be shown as a hyperlink. However, it was also passing the
escaped message to the fix itself, which meant that it no longer matched the
specific pattern expected by the typography detector code which turns its
own error strings into replacement edits.

Change-Id: I25b1509e33145e93f257720daea26c3731a329a2