AndroidSdkData Singleton
This change removes references to SdkManager from Android Studio
and replaces them with references to AndroidSdkData which works
in a singleton-like fashion. It maintains a mapping of files
to the LocalSdk objects which represent them. Since we only
ever have one single SDK in AS, this is effectively a singleton
for Android Studio. The path-to-sdk mapping is left in place to
allow IntelliJ projects to have multiple SDKs.

Change-Id: Iefb39c9926300c7b88cc1fbaa116c852a32e5a6d