Make Nav. Editor support AS's  'master-detail' template
This CL adds 'smarts' to recognize the patterns used in the
master-detail example in the Android Studios templates.

In particular:

1. Respect 'layout aliases' (by parsing refs.xml, in the 'values' dir).

2. Inspect implementation of the onListItemClick()
method, as defined in the ListFragment class of the
v4 support library.

3. Add code to recognize the idiom of having a Fragment
call a method that is implemented in the enclosing activity.
In this case, continue the analysis in the Activity's
implementation of that interface.

Change-Id: I34a6269ae39d41080add9043e8b39958496a19bc