64901: Enable "Filter logcat by selected process" by default
logcat has the option to automatically filter based on the
currently selected process. This is controlled by a user setting,
whose expected behavior was:
 - it should be enabled by default
 - it should be turned off when users explicitly select a filter

However, it was always off by default due since the action of
initializing the list of filters was assumed to be the same as
users explicitly selecting a filter.

This CL fixes the issue and enables it by default by first
initializing the list and then adding a listener to look for
user filter selection events.

Change-Id: I838e1d7e273e6c6ed9bcdb38a56cea36807fa04d