Reuse previous debug session so that watch variables are retained.
While debugging Android apps, the watches panel loses track of
previously watched variables. As a result, users have to manually
add watched variables every time they debug.

The root cause of the issue seems to be that IDEA copies the watches
list to a new debug session only if it can determine that the
previous debug session can be reused, and Android app debug config
wasn't being reused.

The fix is in two parts: The first part is to make sure that when
searching for a previous session, the config can detect that the
same devices are still available to run on (if the launch config
is set to use manual device chooser). The second fix is to make
sure that even while initialization, we use a RemoteDebugRunner
as opposed to a DefaultDebugRunner.

This fixes https://code.google.com/p/android/issues/detail?id=65056

Change-Id: I23b5086bcdfc2df1262b3606e03914071e9a0ce2