Adding support for <icon> tag in templates
The notification template uses an "<icon>" tag which should add a step
to the NewTemplateObjectWizard allowing the user to create a custom
asset for the notification.

This change implements this functionality. In order to do this,
this change also refactors AssetSetStep to allow it to be
included in another wizard and refactors the image generator
functions into their own class; AssetStudioAssetGenerator.

Fixes https://code.google.com/p/android/issues/detail?id=62455

Change-Id: I78883efd09a0ea1471140d9adc83e7a5cb418a38