Prevent importing Eclipse projects of the wrong type
ADT projects and Eclipse projects are not exactly the same; you can
import an ADT project which doesn't have Eclipse metadata files, and
similarly even if a project does have .classpath and .project files,
it might not be an ADT project.

This CL updates our import handler such that if it sees an Eclipse
project which is not an ADT project, it won't fall back to the generic
IntelliJ Eclipse import, it will explain that you should select the
ADT project instead and try to import that.

Change-Id: I4541fa79498928d5863f8831f96ffd29b598b86a