Fix caching of manifest data based on timestamps.
The ManifestInfo class parses and stores information from the manifest
only when it detects that the timestamp of the manifest is newer
than the last seen version.

A recent CL added the ability to parse information from a merged
manifest rather than just the source manifest. When using the merged
manifest, the timestamp check fails because the merged manifest is
parsed using a copy of its contents in memory and not the file.
This implies that changes to the merged manifest weren't being picked
up and a stale copy was being used.

This CL fixes the issue by introducing a new ManifestFile class. This
class encapsulates the knowledge of whether the source of the manifest
is the real manifest file, or a in memory copy of the merged manifest.
Based on this information, it can decide when a manifest needs to be
reparsed - checking the time stamps of the virtual file, or the I/O
file as necessary.

Change-Id: Ib692711b1979f6fa1e96e4eca32e77ab8f67527c