Fixed Gradle daemon shutdown.
There were cases that a Gradle installation could not be found
resulting in a failed attempt to stop all Gradle daemons.

In this CL I added a method to find *any* Gradle installation,
not just the one used by a specific project. Since we stop
daemons by calling the Gradle executable, we can use any
installation.

If a Gradle installation cannot be found before killing daemons
while closing a project, as a last resource, we ask the user
for the path of Gradle installation, and we *store* that path
for future usage.

Change-Id: I9b8ad4d03569a3759dac2e9172ba560a3d16fbf4