Fixed module dependency setup.
The problem:
To fix library dependency setup I unified the 2 code
paths that did this separately. It actually fixed the
problem but created a new one: module dependency was
broken. When importing a project that had inter-module
dependencies, Studio will no longer set up a dependency
to the module, but to the compiled library instead.
Everything would work, but users lost the ability to
navigate to the source code of a module dependency.

The root of the problem was the order in which we were
setting up things. When importing/syncing a project
we were configuring the AndroidGradleFacet last. This
facet has the "Gradle path" of a module (e.g. :app).
When setting up module dependencies, if a module
with a given path was not found, Studio would link to
the compiled artifact. This order never mattered
before, but now that we have a unified way to set up
dependencies, it does.

The solution:
The solution is simply change the order in which the
ProjectDataServices are called. The first one has to
be GradleProjectDataService, so each module can have
a "Gradle path" *before* setting up dependencies.
This order is specified by the "keys" in
AndroidProjectKeys.

In this CL I also moved some "cleanup" code that was
performed by GradleProjectDataService to the new
class ProjectStructureSanitizer. Now
AndroidProjectDataService invokes this code, because
it is executed last.

Change-Id: Ieb0388dc154dfc6ceffc4d9dbdec844bb2c49f2a