Improved Maven dependency lookup
Makes many improvements in the Maven lookup for Project Structure
dependencies. It has a complete rewrite of the class with more
streamlined logic. It uses the new Maven indexing in IntelliJ
for faster and better searches. It has an improved UI that shows
a hardcoded list of very common libraries that users can add with
two clicks. It has some hardcoded search rewrites to make other
searches for common but hard-to-search libraries more successful;
those rewrites were taken from gradleplease.

Bug: 61689
Bug: 60960
Bug: 60959
Bug: 60957

Change-Id: I4b307ef0a9b753b7c4319e0666d8be265255e0fa