Fixed bug where Studio could not recognize different libraries with same name.
The problem:
When setting up library dependencies, we use the artifact's name as the
name of the library. There are libraries that have the same artifact name,
but different qualified package, for example:

    compile 'com.mcxiaoke.volley:library:1.0.0'
    compile 'com.etsy.android.grid:library:1.0.0'

resulting in only one library created, and references to the other one could
not be resolved.

The solution:
We detect if two libraries refer to the same artifact by looking at their
artifacts paths in the file system. If 2 libraries have the same name but
different paths (like in the example above) we create 2 libraries in Studio,
but in the case of our example, we'll have library_1_0_0.xml and
1_library_1_0_0.xml, given that we don't have the Maven coordinate to
deduce the package name.

Fixes
https://code.google.com/p/android/issues/detail?id=65654

Change-Id: Ieaa46b778836a88fd80471f0239437ed7eb1019d