Hides/disables "Gradle invoker" and "Gradle Console" if IDE is not Android Studio.
This CL also enforces usage of JPS if IDE is not Android Studio.

Why?
Android Studio is the only IDE that supports "Gradle invoker" for now. It is
a lightweight alternative to JPS. It is also tied to the "Gradle console".
Since IDEA does not support "Gradle invoker" yet, we need to enforce usage
of JPS for project building. Also, since the "Gradle console" is useless
and confusing without "Gradle invoker" we hide it as well.

Change-Id: I181ede4f602ed7cd7bec9d934f8ee30d7c344bd4