Fixed "Need Sync" notification after project creation/import.
Problem:
When creating or importing a project, opened files get the
notification "Project sync needed" even though a sync just
happened.

In this CL we update the last sync timestamp after the sync
has finished *and* we have a project. Before this change
we tried to update this timestamp on 
GradleImportNotification#onEnd (when sync finished) but we
did not have a Project yet at that moment and the "last
Sync timestamp" was never updated.

Fixes:
https://code.google.com/p/android/issues/detail?id=65844

Change-Id: I23778d974fb9638bc85116f5edd63503e28c884a