Fixed bug where Studio would not generate sources after creating new project.
The problem:
I moved the call to ProjectBuilder#generateSourcesOnly from
ProjectStructureSanitizer to GradleProjectImporter to have more control
on whether we should generate sources after a sync (e.g. we should
not generate sources if the sync was automatically triggered by a
project build.)
This change introduced a bug. When creating a project
ProjectBuilder#generateSourcesOnly is called after sync, but, in the
case of a new project, we haven't got any modules yet, and Studio
cannot figure out which Gradle tasks to invoke, resulting in sources
not being generated at all.

The fix:
Move ProjectBuilder#generateSources back to where it was, because
when invoked there, we know for sure that a project has modules and
it has been properly configured based on the Gradle model.
In this CL, I renamed ProjectStructureSanitizer#cleanup to
PostProjectSyncTasksExecutor#onProjectSetupCompletion to better
reflect that we not just cleaning up a project after a sync, we
are executing necessary tasks that may or may not clean up a project.

Change-Id: Icbb1281168e2c25c06a887acd9a5d16525e9e951