Add main menu to action bar.
This changeset hardcodes the addition of main menu to the action bar.
Also the overflow popup is hardcoded to be shown always. Editor support
still needs to be added to enable users to customise which menu is added
to the layout rendering and whether the overflow menu is shown or not.

Change-Id: Ib08e5313bd7ea5732faf528e67f912791cdba8cc