Fixed updates to editor notifications and Build Variants view.
Problem:
Android Studio provides the following editor editor notifications:

1. When a project sync is in progress
2. When a project sync failed
3. When a project sync may be needed (stale Gradle model)

Problem with #1:
When creating a new project and there is already an open project,
the editors in the existing project will get the notification
"Gradle sync in progress...", even though that sync belongs to the
new project. This is wrong behavior.

Problem with #2:
Sometimes when syncing a project, and even though the sync was
successful, for a second or two, the notification that the sync
failed shows up. The reason for this is that Studio's sync
listener does not wait till the project has been fully set up
after the sync. It thinks the sync has finished before such setup.

Problem with #3:
The "project may need sync" notification shows up even when
editing gradle files in non-Gradle projects.

The root of the problem is how we listen for sync start/end events.
We use a ExternalSystemTaskNotificationListenerAdapter which
automatically is notified when a sync is started/ended. Using this
type of listener has these drawbacks:

1. It does not have any information about which project is being
   synced
2. It does not wait until the project has been fully set up after
   the sync, before letting know editor notifications and the
   "Build Variants" tool window that a sync has ended

As a side effect of #2, the "Build Variants" tool window from time
to time is never loaded and shows up a "Loading" message.

Solution:
This CL introduces the class GradleSyncState which, instead of
being notified automatically when a sync starts and ends, it is
notified manually (since we know exactly when a sync starts
and ends, because we have full control of project sync.)

The benefits of this switch are:
- We can control when to update the IDE UI. Before we were doing
  right after the sync but before the project was set up. Now
  we update the UI after the project has been set up.
- This new class has sync state *per project*, which allows us
  to update the IDE instance of a specific project, instead of
  all IDE instances.

For the problem where we show the notification that a sync may
be needed in non-Gradle projects, now we only check if the
sync is actually needed (before we used to check for YES and
UNSURE in 'isSyncNeeded', now we only check for YES.) YES can
only be returned when dealing with Gradle projects.

Fixes:
https://code.google.com/p/android/issues/detail?id=66009
https://code.google.com/p/android/issues/detail?id=66019

Change-Id: I88eabbfc5ed73dbb3d6a377b6cb1ae1f76e8b472