Add Visibility to Parameter
This change adds a "visibility" attribute to the <parameter> tag
of Template.xml. This attribute should be a Freemarker expression
that evaluates to a Freemarker boolean.
e.g. "minApi gt 6" or "resDir == 'src/main/res'"

The corresponding JComponent will automatically be hidden from the UI.
The JLabel associated can also be hidden automatically if it is
registered with the new "registerLabel" api.

Change-Id: I2009c8337dc246e3e5a88009e9e39ae2ffbe5fbd